feat: 상수 및 함수 분리
