+++ src/app/(home)/_components/DeliveryTimer.tsx
@@ -1,21 +1,7 @@
 "use client";
 
+import formatTime from "@/utils/formatTime";
 import { useEffect, useState } from "react";
-
-function formatTime(ms: number) {
-  const totalSeconds = Math.floor(ms / 1000);
-  const hours = Math.floor(totalSeconds / 3600);
-  const minutes = Math.floor((totalSeconds % 3600) / 60);
-  const seconds = totalSeconds % 60;
-
-  const pad = (n: number) => String(n).padStart(2, "0");
-
-  if (hours > 0) {
-    return `${pad(hours)}:${pad(minutes)}:${pad(seconds)}`;
-  } else {
-    return `${pad(minutes)}:${pad(seconds)}`;
-  }
-}
 
 export default function DeliveryTimer({ user }: { user: null | object }) {
   const [remaining, setRemaining] = useState<number | null>(null);
