+++ src/app/(home)/page.tsx
@@ -1,3 +1,4 @@
+import { CATEGORY_LIST } from "@/constants/category";
 import { cookies } from "next/headers";
 import Image from "next/image";
 import Link from "next/link";
@@ -7,14 +8,6 @@ import Card from "@/components/Card/Card";
 
 import DeliveryTimer from "./_components/DeliveryTimer";
 
-const CATEGORY_LIST = [
-  { name: "문짝", href: "/" },
-  { name: "마감재", href: "/" },
-  { name: "부분장", href: "/" },
-  { name: "부속", href: "/" },
-  { name: "도어락", href: "/" },
-];
-
 async function Page() {
   const cookieStore = await cookies();
   const token = cookieStore.get("token");
@@ -23,7 +16,7 @@ async function Page() {
     redirect("/login");
   }
 
-  const user = null
+  const user = null;
 
   return (
     <div className="mx-5 flex flex-col gap-4 py-9">
@@ -36,7 +29,7 @@ async function Page() {
         </div>
         <Image src="/icons/Headphones.svg" width={24} height={24} alt="문의하기 버튼" />
       </div>
-      <DeliveryTimer user={user}/>
+      <DeliveryTimer user={user} />
       <section className="grid grid-cols-2 gap-2">
         {CATEGORY_LIST.map(item => (
           <Card key={item.name} title={item.name} href={item.href} />
