+++ src/constants/category.ts
@@ -1,0 +1,12 @@
+export type Category = {
+  name: string;
+  href: string;
+};
+
+export const CATEGORY_LIST = [
+  { name: "문짝", href: "/" },
+  { name: "마감재", href: "/" },
+  { name: "부분장", href: "/" },
+  { name: "부속", href: "/" },
+  { name: "", href: "/" },
+];
